Posix dependencies resolved
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk/trdp@9 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ src/common/trdp_mdcom.c
@@ -80,8 +80,8 @@ TRDP_ERR_T  trdp_sendMD (
 TRDP_ERR_T  trdp_rcvMD (
     int         mdSock,
     MD_HEADER_T * *ppPacket,
-    ssize_t     *pSize,
-    uint32_t    *pIPAddr)
+    INT32       *pSize,
+    UINT32      *pIPAddr)
 {
     TRDP_ERR_T err = TRDP_UNKNOWN_ERR;
